Rework for more labels, fewer metrics
While working through the design, it seemed like more potential use
cases needed to know if the metric was for a root policy or not. And if
that is added, then the "total_managed" number could be calculated in
Prometheus.

Additionally, tracking whether the policy is disabled doesn't seem very
useful. In fact, it could confuse calculations that didn't consider it.
Removing metrics on non-active policies seems more intuitive.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>